refactor: use docs api
