--- apps/web/src/lib/constants-client.ts
@@ -1,1 +1,0 @@
-export const PROPERTIES = ["classes", "enums", "variables", "typeAliases", "interfaces", "functions", "namespaces"] as const;
