+++ apps/web/src/lib/constants.ts
@@ -1,3 +1,2 @@
-import { join } from "node:path";
-
-export const DOCS_DIRECTORY = join(process.cwd(), "..", "..", "docs");
+export const BASE_API_ROUTE = "https://docs.snowcrystals.dev/api" as const;
+export const PROPERTIES = ["classes", "enums", "variables", "typeAliases", "interfaces", "functions", "namespaces"] as const;
